requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.3

# Google Generative AI
google-generativeai==0.3.2

# Vector database
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For better CSV handling
openpyxl==3.1.2
